docs(gallery): split plot_geometric_patterns into 5 pattern examples
Flower of Life, Penrose tiling, Islamic geometric pattern, generative
mandala, and fractal forest each become standalone gallery items
with their own thumbnail, so each card on the index renders one
code + one figure.
